# chatgpt-app/app_manifest/openapi.yaml
openapi: 3.1.0
info:
  title: Shopping MCP companion REST surface (OAuth + health)
  version: "0.1.0"
  description: |
    The primary integration surface for ChatGPT in this POC is **MCP over HTTPS** at `{PUBLIC_BASE_URL}/mcp`.
    This OpenAPI document describes the supporting OAuth and health endpoints used alongside MCP.

servers:
  - url: https://YOUR-SUBDOMAIN.ngrok-free.app
    description: Replace with your public base URL (ngrok or other HTTPS tunnel).

paths:
  /health:
    get:
      summary: Liveness probe
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /auth/login:
    get:
      summary: Mock OAuth2 login page
      parameters:
        - name: client_id
          in: query
          schema:
            type: string
        - name: redirect_uri
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: HTML login form

  /auth/callback:
    get:
      summary: OAuth callback — exchanges authorization code for token/session
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [json]
      responses:
        "200":
          description: HTML success page or JSON token payload when format=json

  /auth/status:
    get:
      summary: Session status (optional Bearer)
      parameters:
        - name: access_token
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Authentication status

  /auth/logout:
    post:
      summary: Revoke session
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                access_token:
                  type: string
      responses:
        "200":
          description: Logged out

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer